add tested methods
